% Generated by roxygen2 (4.0.2): do not edit by hand
\name{allIds}
\alias{allIds}
\title{Returns all identifiers and terms of an ontology}
\usage{
allIds(ontologyName, simplify = TRUE)
}
\arguments{
\item{ontologyName}{A \code{character} with the name of a valid ontology name.}

\item{simplify}{A logical indicating whether the S4 \code{Map}
instance should be simplified. Default is \code{TRUE}.}
}
\value{
A named \code{character} if \code{simplify} is \code{TRUE}.
An instance of class \code{Map} otherwise.
}
\description{
This function returns all identifiers and terms available
for given valid ontology. It sends a
\code{getAllTermsFromOntologyRequest} SOAP message
and retrieves and parses the \code{getAllTermsFromOntologyResponse}.
The original interface is
\code{public Map getAllTermsFromOntology(String ontologyName)}.
}
\examples{
allIds("MS", simplify=FALSE)
}
\author{
Laurent Gatto
}
\seealso{
Other ols-queries: \code{\link{isIdObsolete}};
  \code{\link{olsQuery}}; \code{\link{olsVersion}};
  \code{\link{ontologies}}; \code{\link{ontologyLoadDate}};
  \code{\link{ontologyNames}}; \code{\link{rootId}};
  \code{\link{termMetadata}}; \code{\link{termXrefs}};
  \code{\link{term}}
}

