% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.properties}
\alias{.properties}
\title{Constructs the query for all properties from a given ontology}
\usage{
.properties(oid, pagesize = 200)
}
\arguments{
\item{oid}{A character with an ontology or an ontology}

\item{pagesize}{How many results per page to return}
}
\value{
An object of class Terms
}
\description{
Constructs the query for all properties from a given ontology
}
