% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.term}
\alias{.term}
\title{Constructs the query for a single term from a given
    ontology}
\usage{
.term(oid, termid)
}
\arguments{
\item{oid}{A character with an ontology or an ontology}

\item{termid}{A character with a term id}
}
\value{
An object of class Term
}
\description{
Constructs the query for a single term from a given
    ontology
}
