% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{key,mapItem-method}
\alias{key}
\alias{key,Map-method}
\alias{key,mapItem-method}
\title{\code{key} slot accessor.}
\usage{
\S4method{key}{mapItem}(object)

\S4method{key}{Map}(object)
}
\arguments{
\item{object}{An instance of class \code{mapItem}.}

\item{object}{An instance of class \code{Map}.}
}
\value{
A \code{character}.

A \code{character}.
}
\description{
\code{key} slot accessor for the \code{mapItem} instances.

\code{key} slot accessor for the \code{Map} instances.
}
\author{
Laurent Gatto
}

