% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeOntologies}
\alias{makeOntologies}
\title{Makes an Ontologies instance based on the response from
    api/ontologies @return}
\usage{
makeOntologies(pagesize = 150)
}
\arguments{
\item{pagesize}{A numeric indicating the number of elements per
page (default in method is 150).}
}
\value{
An object of class Ontologies
}
\description{
Makes an Ontologies instance based on the response from
    api/ontologies @return
}
