% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeOntology}
\alias{makeOntology}
\title{Makes an Ontology instance based on the response from
    /api/ontologies/{ontology_id}}
\usage{
makeOntology(x)
}
\arguments{
\item{x}{A valid onology prefix}
}
\value{
An object of class Ontology
}
\description{
Makes an Ontology instance based on the response from
    /api/ontologies/{ontology_id}
}
