% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeProperty}
\alias{makeProperty}
\title{Makes a Property instance based on the response from
    /api/ontologies/{ontology}/terms/{iri}}
\usage{
makeProperty(x)
}
\arguments{
\item{x}{The content from the response}
}
\value{
An object of class Property
}
\description{
Makes a Property instance based on the response from
    /api/ontologies/{ontology}/terms/{iri}
}
