% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeTerm}
\alias{makeTerm}
\title{Makes a Term instance based on the response from
    /api/ontologies/{ontology}/terms/{iri}}
\usage{
makeTerm(x)
}
\arguments{
\item{x}{The content from the response}
}
\value{
An object of class Term
}
\description{
Makes a Term instance based on the response from
    /api/ontologies/{ontology}/terms/{iri}
}
