% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{map,ANY-method}
\alias{map}
\alias{map,ANY-method}
\title{\code{Map} slot accessor.}
\usage{
\S4method{map}{ANY}(object)
}
\arguments{
\item{from}{An S4 class produced by an OLS return message.}
}
\value{
A instance of class \code{Map}.
}
\description{
Accessor for the \code{Map} data of the OLS
return messages converted to their respective
S4 classes. The actual data is stored in \code{Map}
slots.
}
\author{
Laurent Gatto
}

