% Generated by roxygen2 (4.0.2): do not edit by hand
\name{olsVersion}
\alias{olsVersion}
\title{Returns the OLS version}
\usage{
olsVersion()
}
\value{
A \code{character} of length 5.
}
\description{
This function returns the Ontology Lookup Webservice
version, build data and author. It sends a
\code{getVersionRequest} SOAP message and retrieves
and parses the \code{getVersionResponse}.
The original interface is \code{public String getVersion()}.
}
\examples{
olsVersion()
}
\author{
Laurent Gatto
}
\seealso{
Other ols-queries: \code{\link{allIds}};
  \code{\link{isIdObsolete}}; \code{\link{olsQuery}};
  \code{\link{ontologies}}; \code{\link{ontologyLoadDate}};
  \code{\link{ontologyNames}}; \code{\link{rootId}};
  \code{\link{termMetadata}}; \code{\link{termXrefs}};
  \code{\link{term}}
}

