% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontologies.R
\name{Ontologies}
\alias{Ontologies}
\alias{Ontology}
\alias{olsLinks}
\alias{olsLinks,Ontology}
\alias{olsConfig}
\alias{olsConfig,Ontology}
\alias{olsVersion}
\alias{olsVersion,character}
\alias{olsVersion,Ontology}
\alias{olsVersion,Ontologies}
\alias{olsLoaded}
\alias{olsLoaded,character}
\alias{olsLoaded,Ontology}
\alias{olsLoaded,Ontologies}
\alias{olsUpdated}
\alias{olsUpdated,character}
\alias{olsUpdated,Ontology}
\alias{olsUpdated,Ontologies}
\alias{olsStatus}
\alias{olsStatus,character}
\alias{olsStatus,Ontology}
\alias{olsStatus,Ontologies}
\alias{olsPrefix}
\alias{olsPrefix,character}
\alias{olsPrefix,Ontology}
\alias{olsPrefix,Ontologies}
\alias{olsDesc}
\alias{olsDesc,character}
\alias{olsDesc,Ontology}
\alias{olsDesc,Ontologies}
\alias{olsTitle}
\alias{olsTitle,character}
\alias{olsTitle,Ontology}
\alias{olsTitle,Ontologies}
\alias{olsNamespace}
\alias{olsNamespace,character}
\alias{olsNamespace,Ontology}
\alias{olsNamespace,Ontologies}
\alias{ontologyUrl}
\alias{ontologyUrl,character}
\alias{ontologyUrl,Ontology}
\alias{as.data.frame.Ontologies}
\alias{Ontologies,missing-method}
\alias{Ontology,character-method}
\alias{Ontology,Ontology-method}
\alias{show,Ontology-method}
\alias{show,Ontologies-method}
\alias{olsVersion,character-method}
\alias{olsVersion,Ontology-method}
\alias{olsVersion,Ontologies-method}
\alias{olsLoaded,character-method}
\alias{olsLoaded,Ontology-method}
\alias{olsLoaded,Ontologies-method}
\alias{olsLinks,Ontology-method}
\alias{olsConfig,Ontology-method}
\alias{olsUpdated,character-method}
\alias{olsUpdated,Ontology-method}
\alias{olsUpdated,Ontologies-method}
\alias{olsPrefix,character-method}
\alias{olsPrefix,Ontology-method}
\alias{olsPrefix,Ontologies-method}
\alias{olsDesc,character-method}
\alias{olsDesc,Ontology-method}
\alias{olsDesc,Ontologies-method}
\alias{olsTitle,character-method}
\alias{olsTitle,Ontology-method}
\alias{olsTitle,Ontologies-method}
\alias{olsStatus,character-method}
\alias{olsStatus,Ontology-method}
\alias{olsStatus,Ontologies-method}
\alias{olsNamespace,character-method}
\alias{olsNamespace,Ontology-method}
\alias{olsNamespace,Ontologies-method}
\alias{ontologyUrl,character-method}
\alias{ontologyUrl,Ontology-method}
\alias{lapply,Ontologies-method}
\alias{[,Ontologies-method}
\alias{[[,Ontologies-method}
\alias{length,Ontologies-method}
\title{Ontologies}
\usage{
\S4method{Ontologies}{missing}(object)

\S4method{Ontology}{character}(object)

\S4method{Ontology}{Ontology}(object)

\S4method{show}{Ontology}(object)

\S4method{show}{Ontologies}(object)

\S4method{olsVersion}{character}(object)

\S4method{olsVersion}{Ontology}(object)

\S4method{olsVersion}{Ontologies}(object)

\S4method{olsLoaded}{character}(object)

\S4method{olsLoaded}{Ontology}(object)

\S4method{olsLoaded}{Ontologies}(object)

\S4method{olsLinks}{Ontology}(object)

\S4method{olsConfig}{Ontology}(object)

\S4method{olsUpdated}{character}(object)

\S4method{olsUpdated}{Ontology}(object)

\S4method{olsUpdated}{Ontologies}(object)

\S4method{olsPrefix}{character}(object)

\S4method{olsPrefix}{Ontology}(object)

\S4method{olsPrefix}{Ontologies}(object)

\S4method{olsDesc}{character}(object)

\S4method{olsDesc}{Ontology}(object)

\S4method{olsDesc}{Ontologies}(object)

\S4method{olsTitle}{character}(object)

\S4method{olsTitle}{Ontology}(object)

\S4method{olsTitle}{Ontologies}(object)

\S4method{olsStatus}{character}(object)

\S4method{olsStatus}{Ontology}(object)

\S4method{olsStatus}{Ontologies}(object)

\S4method{olsNamespace}{character}(object)

\S4method{olsNamespace}{Ontology}(object)

\S4method{olsNamespace}{Ontologies}(object)

\S4method{ontologyUrl}{character}(object)

\S4method{ontologyUrl}{Ontology}(object)

\S4method{lapply}{Ontologies}(X, FUN, ...)

\S4method{[}{Ontologies}(x, i, j = "missing", drop = "missing")

\S4method{[[}{Ontologies}(x, i, j = "missing", drop = "missing")

\S4method{length}{Ontologies}(x)
}
\arguments{
\item{object}{an instance of class `Ontologies` or `Ontology`. For
some functions, a ontology identifier is applicable.}

\item{X}{`Ontologies` object.}

\item{FUN}{a `function` to be applied to each `Ontology` element
of `X`.}

\item{...}{additional arguments passed to `FUN`.}

\item{x}{an `Ontologies` object.}

\item{i}{index of elecements to subset.}

\item{j}{ignored.}

\item{drop}{ignored.}
}
\description{
The rols package provides an interface to PRIDE's Ontology Lookup
Servive (OLS) and can be used to query one or multiple ontologies,
stored as `Ontology` and `Ontologies` instances, and containing
various information as provided by OLS.
}
\details{
Ontologies are referred to by their namespace, which is lower
case: the Gene Onology is "go", the Mass spectrometry ontology is
"ms", etc. The ontologies also have prefixes, which are upper
case: the Gene Onology prefix "GO", the Mass spectrometry ontology
prefix "MS". One exception to this rule is the Drosophila
Phenotype Ontology, whose namespace and prefix are "dpo" and
"FBcv" respectively (there might be more). This is particularly
confusing as the FlyBase Controlled Vocabulary has "fbcv" and
"FBcv" as namespace and prefix respectively.

When using a character to initialise an ontology or query a term,
"fbcv" (this is case insensitive) will refer to the the FlyBase
Controlled Vocabulary. The the Drosophila Phenotype Ontology will
have to be referred as "dpo" (also case insensitive).
}
\section{Constructors}{


Objects can be created in multiple ways. The [Ontologies()]
function will initialise all available ontolgies as an
`Ontologies` object, while a call to [Ontology()] with an ontology
namespace or prefix as argument will initialise the ontology of
interest as an `Ontology` instance.

`Ontolgies` instances can be subset with `[` and `[[` (using their
namespace, see Details) and iterated over with
`lapply`. `Ontolgies` can be converted into a simple `data.frame`
containing the ontology prefixes, namespaces and titles using
`as(., "data.frame")`. `Ontologies` can also be coerced to lists
of `Ontology` ojects with `as(., "list")`.
}

\section{Accessors}{


- `olsDesc(object = "Ontology")` returns the description of an
  ontology. Also works for `Ontologies` objects and a `character`
  describing an ontology namespace or prefix (see Details).

- `olsPrefix(object = "Ontology")` retruns the prefix of an
  ontology. Also works for `Ontologies` objects and a `character`
  describing an ontology namespace or prefix (see Details).

- `olsVersion(object = "Ontology")` returns the version of the
  ontology. Also works with an a `character` defining an ontology
  namespace or prefix (see Details) or an object of class
  `Ontologies`, in which case it returns a list of versions.

- `olsLoaded(object = "Ontology")` returns the loading date of the
  ontology. Also works with a `character` containing the ontology
  namespace or prefix (see Details) or an object of class
  `Ontologies`.

- `olsUpdated(object = "Ontology")` returns the update date of the
  ontology. Also works with a `character` containing the ontology
  namespace or prefix (see Details) or an object of class
  `Ontologies`.

- `olsStatus(object = "Ontology")` returns the status of the
   ontology. Also works with a `character` containing the ontology
   namespace or prefix (see Details) or an object of class
   `Ontologies`.

- `olsTitle(object = "Ontology")` returns the title of an
   ontology. Also works with a `character` containing the ontology
   namespace or prefix (see Details) or an object of class
   `Ontologies`.

- `olsNamespace(object = "Ontology")` returns the namespace of an
  ontology. Also works with a `character` containing the ontology
  namespace or prefix (see Details) or an object of class
  `Ontologies`.

- `olsLinks(object = "Ontology")` returns a named `character` with
  hyperlink to the ontology itself, and other associated concepts
  such as its terms.

- `olsConfig(object = "Ontology")` returns a list of additional
   unstructured, partly redundant information about the ontology.

- `ontologyUrl(object = "Ontology") return the hyperlink to the
  ontology itself. It can also be used with a `character` defining
  the namespace or prefix of an ontology, in which case it is
  created from the base OLS API URL.
}

\section{Ontology terms}{


Once an ontology has been created an an `Ontology` instance, all
its terms can be requested using the `Terms()` constructor. See
[Terms()] for details.
}

\examples{

#############################
## All ontologies
(onts <- Ontologies())

#############################
## Alzheimer's Disease Ontology (ADO)
## 1. From the ontologies object
(ado1 <- onts[['ado']])
## 2. Create from its namespace
(ado2 <- Ontology('ado')) ## also works with ADO

all.equal(ado1, ado2)

olsVersion(ado1)
olsPrefix(ado1)
olsNamespace(ado1)
olsTitle(ado1)
olsDesc(ado1)
olsLinks(ado1)
str(olsConfig(ado1))
}
\references{
- OLS3 API (the OLS4 API should function identically to the OLS3):
  <http://www.ebi.ac.uk/ols4/ols3help>

- REST API for OLS: <https://www.ebi.ac.uk/ols4/swagger-ui/index.html>
}
\author{
Laurent Gatto
}
