\name{ontologyNames}
\alias{ontologyNames}
\title{Returns all ontologyNames}
\usage{
  ontologyNames()
}
\value{
  A \code{charcter} with all ontology names.
}
\description{
  Returns the names of the OLS ontologies.
}
\examples{
head(ontologyNames())
}
\author{
  Laurent Gatto
}
\seealso{
  Other ols-queries: \code{\link{allIds}},
  \code{\link{isIdObsolete}}, \code{\link{olsQuery}},
  \code{\link{olsVersion}}, \code{\link{ontologies}},
  \code{\link{ontologyLoadDate}}, \code{\link{rootId}},
  \code{\link{term}}, \code{\link{termMetadata}},
  \code{\link{termXrefs}}
}

