\docType{package}
\name{rols-package}
\alias{rols}
\alias{rols-package}
\title{R OLS interface}
\description{
  An R interface to the Ontology Lookup Service
}
\details{
  The Ontology Lookup Service (OLS) is a spin-off of the
  PRIDE project. It provides a web service interface to
  query multiple ontologies from a single location with a
  unified output format. This package allows to query the
  OLS from within R.
}
\author{
  Laurent Gatto \email{lg390@cam.ac.uk}
}
\references{
  http://www.ebi.ac.uk/ontology-lookup/
}
\keyword{package}

