\docType{methods}
\name{show,Map-method}
\alias{show}
\alias{show?Map}
\alias{show,Map-method}
\title{\code{Map} show method}
\arguments{
  \item{object}{An \code{Map} instance.}
}
\value{
  Returns an invisible 'NULL'. This function is used for
  its side-effect of printing a textual description of
  \code{object}.
}
\description{
  show method for \code{Map} instances
}
\author{
  Laurent Gatto
}

