\docType{methods}
\name{value,mapItem-method}
\alias{value}
\alias{value,mapItem-method}
\alias{value,Map-method}
\title{\code{value} slot accessor.}
\arguments{
  \item{object}{An instance of class \code{mapItem}.}

  \item{object}{An instance of class \code{Map}.}
}
\value{
  A \code{character}.

  A \code{character}.
}
\description{
  \code{value} slot accessor for the \code{mapItem}
  instances.

  \code{value} slot accessor for the \code{Map} instances.
}
\author{
  Laurent Gatto
}

