\docType{methods}
\name{key,mapItem-method}
\alias{key}
\alias{key,mapItem-method}
\alias{key,Map-method}
\title{\code{key} slot accessor.}
\arguments{
  \item{object}{An instance of class \code{mapItem}.}

  \item{object}{An instance of class \code{Map}.}
}
\value{
  A \code{character}.

  A \code{character}.
}
\description{
  \code{key} slot accessor for the \code{mapItem}
  instances.

  \code{key} slot accessor for the \code{Map} instances.
}
\author{
  Laurent Gatto
}

