% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Properties.R
\name{Properties}
\alias{Properties}
\title{Term Properties}
\description{
Properties (relationships) between terms can be queries for
complete [Ontology()] objects and [Term()]/[Terms()] instances,
and the results are stored as objects of class `Property` or
`Properties`.
}
\examples{

## Term properties
trm <- Term("uberon", "UBERON:0002107")
trm
Properties(trm)

## Ontology properties
Properties('ado')
}
