% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,Map-method}
\alias{show}
\alias{show,Map-method}
\alias{show?Map}
\title{\code{Map} show method}
\usage{
\S4method{show}{Map}(object)
}
\arguments{
\item{object}{An \code{Map} instance.}
}
\value{
Returns an invisible 'NULL'. This function is used
for its side-effect of printing a textual description
of \code{object}.
}
\description{
show method for \code{Map} instances
}
\author{
Laurent Gatto
}

