% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{value,mapItem-method}
\alias{value}
\alias{value,Map-method}
\alias{value,mapItem-method}
\title{\code{value} slot accessor.}
\usage{
\S4method{value}{mapItem}(object)

\S4method{value}{Map}(object)
}
\arguments{
\item{object}{An instance of class \code{mapItem}.}

\item{object}{An instance of class \code{Map}.}
}
\value{
A \code{character}.

A \code{character}.
}
\description{
\code{value} slot accessor for the \code{mapItem} instances.

\code{value} slot accessor for the \code{Map} instances.
}
\author{
Laurent Gatto
}

